@@ -2,7 +2,7 @@
 name = "spdlog-internal"
 version = "0.1.0"
 edition = "2021"
-rust-version = "1.60"
+rust-version = "1.61"
 description = "Internal private common code for crate \"spdlog-rs\""
 repository = "https://github.com/SpriteOvO/spdlog-rs"
 license = "MIT OR Apache-2.0"
