@@ -2,7 +2,7 @@
 name = "spdlog-macros"
 version = "0.2.0"
 edition = "2021"
-rust-version = "1.60"
+rust-version = "1.61"
 description = "Macros implementation of crate \"spdlog-rs\""
 repository = "https://github.com/SpriteOvO/spdlog-rs"
 license = "MIT OR Apache-2.0"
