@@ -2,7 +2,7 @@
 name = "spdlog-rs"
 version = "0.4.3"
 edition = "2021"
-rust-version = "1.60"
+rust-version = "1.61"
 description = "Fast, highly configurable Rust logging crate, inspired by the C++ logging library spdlog"
 repository = "https://github.com/SpriteOvO/spdlog-rs"
 license = "MIT OR Apache-2.0"
@@ -34,10 +34,12 @@ release-level-trace    = []
 
 source-location = []
 native = []
-libsystemd = ["libsystemd-sys"]
-multi-thread = ["crossbeam"]
-runtime-pattern = ["spdlog-internal"]
+libsystemd = ["dep:libsystemd-sys"]
+multi-thread = ["dep:crossbeam"]
+runtime-pattern = ["dep:spdlog-internal"]
+serde = ["dep:serde", "value-bag/serde"]
 serde_json = ["serde", "dep:serde_json"]
+sval = ["value-bag/sval"]
 
 [dependencies]
 arc-swap = "1.5.1"
@@ -49,14 +51,14 @@ dyn-clone = "1.0.14"
 flexible-string = { version = "0.1.0", optional = true }
 if_chain = "1.0.2"
 is-terminal = "0.4"
-log = { version = "0.4.8", optional = true }
+log = { version = "0.4.21", optional = true, features = ["kv"] }
 once_cell = "1.16.0"
 serde = { version = "1.0.163", optional = true, features = ["derive"] }
 serde_json = { version = "1.0.120", optional = true }
 spdlog-internal = { version = "=0.1.0", path = "../spdlog-internal", optional = true }
 spdlog-macros = { version = "=0.2.0", path = "../spdlog-macros" }
-spin = "0.9.8"
 thiserror = "1.0.37"
+value-bag = { version = "1.11.1", features = ["owned", "inline-i128"] }
 
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3.9", features = ["consoleapi", "debugapi", "handleapi", "processenv", "processthreadsapi", "winbase", "wincon"] }
@@ -73,6 +75,8 @@ crossbeam = "0.8.2"
 regex = "1.7.0"
 paste = "1.0.14"
 trybuild = "1.0.90"
+sval = "2.14.0"
+sval_derive = "2.14.0"
 
 # [bench-dependencies]
 log = "=0.4.22"
@@ -119,6 +123,9 @@ required-features = ["runtime-pattern", "serde_json"]
 name = "spdlog_rs_log_crate_proxy"
 path = "benches/spdlog-rs/log_crate_proxy.rs"
 required-features = ["log"]
+[[bench]]
+name = "spdlog_rs_kv"
+path = "benches/spdlog-rs/kv.rs"
 [[bench]]
 name = "fast_log"
 path = "benches/fast_log/main.rs"
